Move handlers to own folder
