+++ lib/handlers/authenticate.js
@@ -2,9 +2,9 @@
 
 const joi = require('joi');
 const boom = require('boom');
-const jwt = require('./jwt_signer');
-const userDb = require('./userDb');
-const errorSchema = require('./error_schema');
+const jwt = require('../jwt_signer');
+const userDb = require('../userDb');
+const errorSchema = require('../error_schema');
 
 const userCredentialsSchema = joi.object({
   id: joi.string().required().description('User identifier'),
