+++ lib/handlers/users_{id}.js
@@ -2,8 +2,8 @@
 
 const joi = require('joi');
 const boom = require('boom');
-const userDb = require('./userDb');
-const errorSchema = require('./error_schema');
+const userDb = require('../userDb');
+const errorSchema = require('../error_schema');
 
 const userSchema = joi.object({
   id: joi.string().required().description('User identifier.'),
@@ -43,5 +43,5 @@ exports.register = function (server, options, next) {
 };
 
 exports.register.attributes = {
-  name: 'users'
+  name: 'users_{id}'
 };
