+++ lib/server.js
@@ -6,8 +6,8 @@ const pkg = require('../package.json');
 const jwtScheme = require('hapi-auth-jwt2');
 
 const jwtSigner = require('./jwt_signer');
-const users = require('./users');
-const authenticate = require('./authenticate');
+const users = require('./handlers/users_{id}');
+const authenticate = require('./handlers/authenticate');
 
 const plugins = [{
   register: users
